pyproject.toml
[tool.poetry]
name = "cell_quests"
version = "0.1.0"
description = ""
authors = ["kszinhu <kszinhu@kszinhu.dev.br>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12.3"
dearpygui = "^1.11.1"
mysql-connector-python = "^8.4.0"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
